Cleanup and refactor ntumc_util * Refactored expandlem() and pos2win() for readability * Refactored tbox() with template() to generate HTML clearly and injectproofly * Removed unused imports (re, sqlite3, collections, operator) * PEP8-tified code using pylint (module-level names for constants untouched)
Some of the remaining code is likely to behave unexpectedly, for example `wndb`
defines a relative path that won't make sense during runtime when said path
will be relative to the main thread's path (`os.path.realpath(__file__)` is a
likely solution). These problematic areas will be addressed if I find that they
are used in other parts of the code.